+++ changelog/15561.txt
@@ -1,0 +1,3 @@
+```release-note:improvement
+ssh: Addition of an endpoint `ssh/issue/:role` to allow the creation of signed key pairs
+```
