+++ website/content/api-docs/secret/ssh.mdx
@@ -873,3 +873,88 @@ $ curl \
   "auth": null
 }
 ```
+
+## Generate Certificate and Key
+
+This endpoint generates a new set of credentials (private key and certificate)
+based on the role named in the endpoint.
+
+~> **Note**: The private key is _not_ stored. If you do not save the private
+   key from the response, you will need to request a new certificate.
+
+| Method | Path               |
+| :----- | :----------------  |
+| `POST` | `/ssh/issue/:name` |
+
+### Parameters
+
+- `name` `(string: <required>)` – Specifies the name of the role to create the
+  certificate against. This is part of the request URL.
+
+- `key_type` `(string: "rsa")` – Specifies the desired key type; must be `rsa`, `ed25519`
+  or `ec`.
+
+- `key_bits` `(int: 0)` – Specifies the number of bits to use for the
+  generated keys. Allowed values are 0 (universal default); with
+  `key_type=rsa`, allowed values are: 2048 (default), 3072, or
+  4096; with `key_type=ec`, allowed values are: 224, 256 (default),
+  384, or 521; ignored with `key_type=ed25519`.
+
+- `ttl` `(string: "")` – Specifies the Requested Time To Live. Cannot be greater
+  than the role's `max_ttl` value. If not provided, the role's `ttl` value will
+  be used. Note that the role values default to system values if not explicitly
+  set.
+
+- `valid_principals` `(string: "")` – Specifies valid principals, either
+  usernames or hostnames, that the certificate should be signed for.
+
+- `cert_type` `(string: "user")` – Specifies the type of certificate to be
+  created; either "user" or "host".
+
+- `key_id` `(string: "")` – Specifies the key id that the created certificate
+  should have. If not specified, the display name of the token will be used.
+
+- `critical_options` `(map<string|string>: "")` – Specifies a map of the
+  critical options that the certificate should be signed for. Defaults to none.
+
+- `extensions` `(map<string|string>: "")` – Specifies a map of the extensions
+  that the certificate should be signed for. Defaults to none.
+
+### Sample Payload
+
+```json
+{
+  "key_type": "rsa",
+  "key_bits":  2048
+}
+```
+
+### Sample Request
+
+```shell-session
+$ curl \
+    --header "X-Vault-Token: ..." \
+    --request POST \
+    --data @payload.json \
+    http://127.0.0.1:8200/v1/ssh/issue/my-role
+```
+
+### Sample Response
+
+```json
+{
+  "request_id": "94fd1102-08a1-c207-0e3e-657e8f80c09e",
+  "lease_id": "",
+  "renewable": false,
+  "lease_duration": 0,
+  "data": {
+    "serial_number": "1e965817eb12a511",
+    "signed_key": "ssh-rsa-cert-v01@openssh.com AAAAHHN...\n",
+    "private_key": "-----BEGIN RSA PRIVATE KEY-----\nMIIEpQIBAAKCAQEAwer03vkQrPV+wWpbisJJv2CKqHmMz+Ej0ctLbhpOmR2CY9S9\n...\nQN351pgTphi6nlCkGPzkDuwvtxSxiCWXQcaxrHAL7MiJpPzkIBq1\n-----END RSA PRIVATE KEY-----\n",
+    "private_key_type": "rsa"
+  },
+  "wrap_info": null,
+  "warnings": null,
+  "auth": null
+}
+```
